pyproject.toml
[tool.poetry]
name = "telephoto-updater"
version = "0.1.0"
description = "bot for autoupdating profile photo"
authors = ["donBarbos <donbarbos@proton.me>"]

[tool.poetry.dependencies]
python = "^3.8"
Telethon = "^1.25.4"
python-dotenv = "^0.21.0"
loguru = "^0.6.0"
aiohttp = "^3.8.3"
aiofiles = "^22.1.0"
colorama = "^0.4.6"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
flake8-bugbear = "^22.10.27"
flake8-polyfill = "^1.0.2"
flake8-quotes = "^3.3.1"
flake8-variables-names = "^0.0.5"
flake8-black = "^0.3.3"
pyflakes = "^2.5.0"
pep8-naming = "^0.13.2"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 79
target_version = ["py37", "py38", "py39"]
multi-line-output = 3
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | \.env
  | logs
  | photos
)/
'''
